requirements.txt
logging
fastapi
pydantic
psycopg2-binary
